config.py
import os
from dotenv import load_dotenv

load_dotenv()
DATA_PATH = 'data/'
MODEL_NAME = 'sentence-transformers/all-MiniLM-L6-v2'
INDEX_NAME = 'chatbot'
PINECONE_API_KEY = os.getenv("PINECONE_API_KEY")

if not PINECONE_API_KEY:
    raise ValueError("❌ Missing PINECONE_API_KEY in .env file")


# Prompt Template
PROMPT_TEMPLATE = """
You are a financial analysis assistant. Use the following context extracted from financial documents
(Income Statements, Balance Sheets, Cash Flow Statements, or related reports) to answer the user’s question.  

Context:
{context}

Question:
{question}

Instructions:
- Focus ONLY on the financial information available in the context.
- If the user asks about revenue, expenses, profit, cash flow, or other metrics, extract the exact numbers if present.
- If a value is not explicitly available in the context, respond with:
  "The document does not provide this information."
- Do NOT make assumptions or generate financial data that is not in the context.
- Keep answers concise, clear, and in plain language.
- If the user asks a follow-up question, maintain conversational consistency.

💡 Answer:
"""